refactor(backend): Express housekeeping — imports, error handler, entrypoint split, NOT NULL
- Normalize express imports to named `{ Router }` across all route files
  (was a mix of default import, `expressLib` alias, and named import)
- Extract `registerStrategies(db)` from `createAuthRouter` so Passport
  strategy registration is an explicit call in `createApp` next to
  `passport.initialize()`, not a hidden side effect inside a factory
- Add global error-handling middleware to `createApp` that returns
  `{ error: "Internal server error" }` with a 500 status
- Split server.ts into a pure app factory (server.ts) and a dedicated
  entrypoint (index.ts); update dev/start scripts accordingly
- Add NOT NULL to `jobs.user_id` in db.ts schema (live DB already migrated)
