+++ backend/package.json
@@ -4,8 +4,8 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "dev": "tsx watch server.ts",
-    "start": "tsx server.ts",
+    "dev": "tsx watch index.ts",
+    "start": "tsx index.ts",
     "test": "vitest run --coverage",
     "tsc": "tsc --noEmit",
     "deploy": "bash ./scripts/deploy-backend.sh"
